+++ src/main/java/br/edu/unifio/odonto/desapego/common/response/PagedResponse.java
@@ -1,0 +1,31 @@
+package br.edu.unifio.odonto.desapego.common.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PagedResponse<T> {
+
+    private List<T> content;
+    private long totalElements;
+    private int totalPages;
+    private int size;
+    private int number;
+
+    public static <T> PagedResponse<T> of(List<T> content, long totalElements, int totalPages, int size, int number) {
+        return PagedResponse.<T>builder()
+                .content(content)
+                .totalElements(totalElements)
+                .totalPages(totalPages)
+                .size(size)
+                .number(number)
+                .build();
+    }
+}