+++ src/main/java/br/edu/unifio/odonto/desapego/listing/domain/ListingStatus.java
@@ -1,0 +1,7 @@
+package br.edu.unifio.odonto.desapego.listing.domain;
+
+public enum ListingStatus {
+    ACTIVE,
+    SOLD,
+    RESERVED
+}